creating a ConsumerOffsetsDao interface
