+++ kafka-rest/src/main/java/io/confluent/kafkarest/controllers/ConsumerGroupLagManagerImpl.java
@@ -43,7 +43,8 @@ final class ConsumerGroupLagManagerImpl implements ConsumerGroupLagManager {
       String consumerGroupId
   ) {
     try {
-      ConsumerGroupLag lag = consumerOffsetsDao.getConsumerGroupOffsets(clusterId, consumerGroupId, IsolationLevel.READ_COMMITTED);
+      ConsumerGroupLag lag = consumerOffsetsDao
+          .getConsumerGroupOffsets(clusterId, consumerGroupId, IsolationLevel.READ_COMMITTED);
       return CompletableFuture.completedFuture(Optional.ofNullable(lag));
     } catch (ArithmeticException e) {
       // log.warn("lag exceeds Integer.MAX_VALUE", e);
